+++ packages/envsync-enterprise/src/index.ts
@@ -1,1 +1,2 @@
 export { enterpriseManagementModules } from "./management-modules";
+export { startEnterpriseSyncWorker, startLicenseHeartbeat } from "./background";